ENH: better types all around

+++ Wrapping/vtkParse.h
@@ -48,30 +48,33 @@ typedef enum _parse_item_t
  */
 typedef enum _parse_access_t
 {
-  VTK_ACCESS_PUBLIC = 0,
+  VTK_ACCESS_PUBLIC    = 0,
   VTK_ACCESS_PROTECTED = 1,
-  VTK_ACCESS_PRIVATE = 2
+  VTK_ACCESS_PRIVATE   = 2
 } parse_access_t;
+
 
 struct _TemplateArgs;
 struct _FunctionInfo;
 
 /**
- * TemplateArgs holds template definitions
+ * TemplateArgs holds one template arg
  */
-
 typedef struct _TemplateArg
 {
-  int   Type;         /* is zero for "typename", "class", "template" */
-  char *Class;        /* class name for type */
-  char *Name;         /* name of template arg */
-  char *Value;        /* default value */
+  unsigned int  Type;  /* is zero for "typename", "class", "template" */
+  char         *Class; /* class name for type */
+  char         *Name;  /* name of template arg */
+  char         *Value; /* default value */
   struct _TemplateArgs *Template; /* for templated template args */
 } TemplateArg;
 
+/**
+ * TemplateArgs holds template definitions
+ */
 typedef struct _TemplateArgs
 {
-  int NumberOfArguments;
+  unsigned long NumberOfArguments;
   TemplateArg **Arguments;
 } TemplateArgs;
 
@@ -80,10 +83,10 @@ typedef struct _TemplateArgs
  */
 typedef struct _ItemInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;
-  char *Comment;
+  char          *Name;
+  char          *Comment;
 } ItemInfo;
 
 /**
@@ -96,19 +99,19 @@ typedef struct _ItemInfo
  */
 typedef struct _ValueInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;
-  char *Comment;
-  char *Value;               /* for vars or default arg values */ 
-  int   Type;                /* as defined in vtkParseType.h   */
-  char *Class;               /* classname for type */
-  int   Count;               /* total number of values, if known */
-  int   NumberOfDimensions;  /* dimensionality for arrays */
-  char **Dimensions;         /* dimensions for arrays */
-  struct _FunctionInfo *Function;    /* for function pointer values */
-  int   IsStatic;            /* for class variables only */
-  int   IsEnum;              /* for constants only */
+  char          *Name;
+  char          *Comment;
+  char          *Value;      /* for vars or default arg values */
+  unsigned int   Type;       /* as defined in vtkParseType.h   */
+  char          *Class;      /* classname for type */
+  unsigned long  Count;      /* total number of values, if known */
+  unsigned long  NumberOfDimensions; /* dimensionality for arrays */
+  char         **Dimensions; /* dimensions for arrays */
+  struct _FunctionInfo *Function;  /* for function pointer values */
+  int            IsStatic;   /* for class variables only */
+  int            IsEnum;     /* for constants only */
 } ValueInfo;
 
 /**
@@ -116,32 +119,32 @@ typedef struct _ValueInfo
  */
 typedef struct _FunctionInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;
-  char *Comment;
-  char *Signature;           /* function signature as text */
-  TemplateArgs *Template;    /* template args, or NULL */
-  int   NumberOfArguments;
-  ValueInfo **Arguments;
-  ValueInfo *ReturnValue;    /* NULL for constructors and destructors */
-  int   IsOperator;
-  int   IsVariadic;
-  int   IsLegacy;            /* marked as a legacy method or function */
-  int   IsStatic;            /* methods only */
-  int   IsVirtual;           /* methods only */
-  int   IsPureVirtual;       /* methods only */
-  int   IsConst;             /* methods only */
-  int   ArgTypes[MAX_ARGS];  /* legacy */
-  char *ArgClasses[MAX_ARGS];/* legacy */
-  int   ArgCounts[MAX_ARGS]; /* legacy */
-  int   ReturnType;          /* legacy */
-  char *ReturnClass;         /* legacy */
-  int   HaveHint;            /* legacy */
-  int   HintSize;            /* legacy */
-  int   ArrayFailure;        /* legacy */
-  int   IsPublic;            /* legacy */
-  int   IsProtected;         /* legacy */
+  char          *Name;
+  char          *Comment;
+  char          *Signature;   /* function signature as text */
+  TemplateArgs  *Template;    /* template args, or NULL */
+  unsigned long  NumberOfArguments;
+  ValueInfo    **Arguments;
+  ValueInfo     *ReturnValue; /* NULL for constructors and destructors */
+  int            IsOperator;
+  int            IsVariadic;
+  int            IsLegacy;    /* marked as a legacy method or function */
+  int            IsStatic;    /* methods only */
+  int            IsVirtual;   /* methods only */
+  int            IsPureVirtual; /* methods only */
+  int            IsConst;     /* methods only */
+  unsigned int   ArgTypes[MAX_ARGS];  /* legacy */
+  char          *ArgClasses[MAX_ARGS];/* legacy */
+  unsigned long  ArgCounts[MAX_ARGS]; /* legacy */
+  unsigned int   ReturnType;  /* legacy */
+  char          *ReturnClass; /* legacy */
+  int            HaveHint;    /* legacy */
+  int            HintSize;    /* legacy */
+  int            ArrayFailure;/* legacy */
+  int            IsPublic;    /* legacy */
+  int            IsProtected; /* legacy */
 } FunctionInfo;
 
 /**
@@ -150,10 +153,10 @@ typedef struct _FunctionInfo
  */
 typedef struct _EnumInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;
-  char *Comment;
+  char          *Name;
+  char          *Comment;
 } EnumInfo;
 
 /**
@@ -161,12 +164,12 @@ typedef struct _EnumInfo
  */
 typedef struct _UnionInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;
-  char *Comment;
-  int NumberOfMembers;
-  ValueInfo **Members;
+  char          *Name;
+  char          *Comment;
+  int            NumberOfMembers;
+  ValueInfo    **Members;
 } UnionInfo;
 
 /**
@@ -174,23 +177,23 @@ typedef struct _UnionInfo
  */
 typedef struct _ClassInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;
-  char *Comment;
-  TemplateArgs *Template;
-  int   NumberOfSuperClasses;
-  char **SuperClasses;
-  int   NumberOfItems;
-  ItemInfo **Items;
-  int   NumberOfFunctions;
+  char          *Name;
+  char          *Comment;
+  TemplateArgs  *Template;
+  unsigned long  NumberOfSuperClasses;
+  char         **SuperClasses;
+  unsigned long  NumberOfItems;
+  ItemInfo     **Items;
+  unsigned long  NumberOfFunctions;
   FunctionInfo **Functions;
-  int   NumberOfConstants;
-  ValueInfo **Constants;
-  int   NumberOfEnums;
-  EnumInfo **Enums;
-  int   IsAbstract;
-  int   HasDelete;
+  unsigned long  NumberOfConstants;
+  ValueInfo    **Constants;
+  unsigned long  NumberOfEnums;
+  EnumInfo     **Enums;
+  int            IsAbstract;
+  int            HasDelete;
 } ClassInfo;
 
 /**
@@ -198,21 +201,21 @@ typedef struct _ClassInfo
  */
 typedef struct _NamespaceInfo
 {
-  parse_item_t ItemType;
+  parse_item_t   ItemType;
   parse_access_t Access;
-  char *Name;  /* NULL for global namespace */
-  char *Comment;
-  int   NumberOfItems;
-  ItemInfo **Items;
-  int   NumberOfClasses;
-  ClassInfo **Classes;
-  int   NumberOfFunctions;
+  char          *Name;  /* NULL for global namespace */
+  char          *Comment;
+  unsigned long  NumberOfItems;
+  ItemInfo     **Items;
+  unsigned long  NumberOfClasses;
+  ClassInfo    **Classes;
+  unsigned long  NumberOfFunctions;
   FunctionInfo **Functions;
-  int   NumberOfConstants;
-  ValueInfo **Constants;
-  int   NumberOfEnums;
-  EnumInfo **Enums;
-  int   NumberOfNamespaces;
+  unsigned long  NumberOfConstants;
+  ValueInfo    **Constants;
+  unsigned long  NumberOfEnums;
+  EnumInfo     **Enums;
+  unsigned long  NumberOfNamespaces;
   struct _NamespaceInfo **Namespaces;
 } NamespaceInfo;
